+++ src/server.js
@@ -18,9 +18,6 @@
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 *
 *
-*
-*
-*
 */
 
 
@@ -99,38 +96,40 @@ webSocketServe.on('connection', (websocket) => {
 
 
 
+function startServer() {
+    const server = http.createServer((request, response) => {
+        response.end('SH exposed HTTP');
+    });
 
+    server.listen(0, '0.0.0.0', () => {
+        const Networkdetail = Object.values(os.networkInterfaces()).flat();
+        const network = Networkdetail.find(details => details.family === 'IPv4' && details.internal === false);
+        const assigned = server.address().port;
+        
+        let address = 'localhost';
+        if (network) {
+            address = network.address;
+        }
 
+        const coralGreen = '\x1b[38;5;167m';
+        const lowerGreen = '\x1b[38;2;180;210;170m';
+        const dimWhite = '\x1b[2m';
+        const reset = '\x1b[0m';
+        const bold = '\x1b[1m';
 
-server
-    .listen(0, '0.0.0.0', () => {
-    const Networkdetail = Object.values(os.networkInterfaces()).flat();
-    const network = Networkdetail.find(details => details.family === 'IPv4' && details.internal === false);
-    
-    const assigned = server.address().port;
-    
-    let address
-    if (network) {
-        address = network.address
-    }
-
-    const coralGreen = '\x1b[38;5;167m'
-    const lowerGreen = '\x1b[38;2;180;210;170m'
-    const dimWhite = '\x1b[2m'
-    const reset = '\x1b[0m'
-    const bold = '\x1b[1m'
-
-console.log(`\n${bold}Shell Exposed HTTP${reset}
+        console.log(`\n${bold}Shell Exposed HTTP${reset}
 ${dimWhite}Status:${coralGreen} Online ${reset}
 ${dimWhite}Port:${coralGreen} ${assigned}${reset}
 
 ${lowerGreen}Local:${reset} http://localhost:${coralGreen}${assigned}${reset}
 ${lowerGreen}Network:${reset} http://${address}:${coralGreen}${assigned}${reset}
 `);
-});
+    });
+}
 
-
-
+if (require.main === module) {
+    startServer();
+}
 
 const cyan = "\x1b[36m";
 const green = "\x1b[32m";
@@ -153,4 +152,7 @@ webapp.use(morgan((tokens, request, response) => {
 webapp.use(
     morgan(`${cyan}:method${recolor} :url ${green}:status${recolor} :res[content-length] - :response-time ms\nIP: :remote-addr\nServer on :referrer`)
 );
+
+
+module.exports = { startServer, getShell };