+++ CHANGELOG.md
@@ -9,7 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Termux terminal support.
-- Termination Feature.
+
+## [1.4.0] - [NOTDATED]
+### Added
+- Command Line Interface
+    - commands
+        - shell: display sheh detected shell
+        - version: print version number
+        - help: show the CLI manual
+        - server: manage childs instances
+            - `--child [number]`
+    - options
+        - `-h, --help`
+        - `-v, --version` 
+
 
 ## [1.3.1] - 2026-05-10
 ### Fixed
@@ -70,9 +83,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-[Unreleased]: https://github.com/waxodium/sheh/compare/v1.3.1...HEAD
+[Unreleased]: https://github.com/waxodium/sheh/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/waxodium/sheh/releases/tag/v1.4.0
 [1.3.0]: https://github.com/waxodium/sheh/releases/tag/v1.3.0
 [1.2.0]: https://github.com/waxodium/sheh/releases/tag/v1.2.0
 [1.1.2]: https://github.com/waxodium/sheh/releases/tag/v1.1.2
 [1.0.2]: https://github.com/waxodium/sheh/releases/tag/v1.0.2
 [1.0.1]: https://github.com/waxodium/sheh/releases/tag/v1.0.1
+
+